LayoutTests/fast/css/color-invalid-contrast-color-function.html

<!DOCTYPE html>
<html>
<head>
<title>Test CSS Color Module 5 contrast-color() parsing</title>
<meta charset="utf-8">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../imported/w3c/web-platform-tests/css/support/parsing-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 10px;
    color: pink;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>

test_invalid_value("background-color", "contrast-color()");
test_invalid_value("background-color", "contrast-color(1)");
test_invalid_value("background-color", "contrast-color(max)");
test_invalid_value("background-color", "contrast-color(max max)");
test_invalid_value("background-color", "contrast-color(max white)");
test_invalid_value("background-color", "contrast-color(white min)");
test_invalid_value("background-color", "contrast-color(white max bar)");

</script>
</body>
</html>
